+++ desktop/php/wallbox.php
@@ -7,7 +7,16 @@ $plugin = plugin::byId('wallbox');
 sendVarToJS('eqType', $plugin->getId());
 $eqLogics = eqLogic::byType($plugin->getId());
 if(count($eqLogics) != 0){
+	$username = config::byKey("username", "wallbox");
+	$password = config::byKey("password", "wallbox");
+	if($username != null || $password != null){
 $chargers = $eqLogics[0]->getChargerList();
+}
+else{
+	log::add('wallbox', 'error', 'Plugin is not configured with username & password wallbox account.' );
+	$charger = [];
+}
+
 }else{$charger = [];}
 ?>
 
@@ -143,7 +152,7 @@ foreach ($chargers as $charger) {
 </label>
 <div class="col-sm-7">
 <div class="input-group">
-<input type="text" class="eqLogicAttr form-control roundedLeft" data-l1key="configuration" data-l2key="autorefresh" placeholder="{{Cliquer sur ? pour afficher l'assistant cron}}"/>
+<input type="text" class="eqLogicAttr form-control roundedLeft" data-l1key="configuration" value="*/5 * * * *" data-l2key="autorefresh" placeholder="{{Cliquer sur ? pour afficher l'assistant cron}}"/>
 <span class="input-group-btn">
 <a class="btn btn-default cursor jeeHelper roundedRight" data-helper="cron" title="Assistant cron">
 <i class="fas fa-question-circle"></i>