+++ plugin_info/configuration.php
@@ -21,7 +21,13 @@ if (!isConnect()) {
   include_file('desktop', '404', 'php');
   die();
 }
+$username = config::byKey("username", "wallbox");
+$password = config::byKey("password", "wallbox");
+if($username == null || $password == null){
 ?>
+<p style="color:red;text-align: center;font-weight: bold;">{{Username & Password are required for plugin.}}</p>
+<?php 
+throw new Exception("A username and password is required for the plugin to work"); } ?>
 <form class="form-horizontal">
   <fieldset>
     <div class="form-group">