+++ docs/fr_FR/changelog.md
@@ -1,10 +1,18 @@
 # Changelog wallbox plugin
 
-# 30/01/2021
+# 10/2022 (1.0.0)
+- Ajout de vérification de configuration du plugin avant de pouvoir effectuer l'ajout d'équipement ou l'activation des CRON
+- Correction sur le format de l'heure de la dernière synchronisation avec la borne UTC => Local Date Time (24h)
+- Ajout de la fonctionnalité de contrôle de charge via le widget
+- Ajout de la fonctionnalité de modification du maximum d'ampère d'un chargeur
+- Ajout de la fonctionnalité de verrouillage via le widget
+- Ajout de l'autorefresh par défaut
+
+# 30/01/2021 (1.0 Beta)
 - Implementation of API call to get chargers of account
 - Changed input to dropdownlist for charger selection
 
-# 28/01/2021
+# 28/01/2021 (1.0 Beta)
 
 - Initialisation du projet
 - Ajouts d'éléments d'informations et de paramètres pour les commandes